requirements.txt
pyyaml
requests
docopt
